# Set: comparer optimization