Same as #10855

The job `Main50` includes previous [PR](https://github.com/dotnet/fsharp/pull/10860)

|                         Method |     Job | BuildConfiguration | Size |         Mean |        Error |       StdDev | Ratio | RatioSD | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------- |-------- |------------------- |----- |-------------:|-------------:|-------------:|------:|--------:|-----:|-------:|------:|------:|----------:|
|                 containsKeyInt |   After |              After |  100 |     21.14 ns |     0.779 ns |     0.121 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|                 containsKeyInt |  Main50 |             Main50 |  100 |     28.95 ns |     1.199 ns |     0.311 ns |  1.37 |    0.02 |    2 |      - |     - |     - |         - |
|                 containsKeyInt | NuGet50 |            NuGet50 |  100 |     46.82 ns |     1.132 ns |     0.175 ns |  2.21 |    0.02 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|             containsKeyIntLike |   After |              After |  100 |     59.67 ns |     1.831 ns |     0.476 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|             containsKeyIntLike |  Main50 |             Main50 |  100 |    248.88 ns |     5.699 ns |     1.480 ns |  4.17 |    0.05 |    2 | 0.0525 |     - |     - |     331 B |
|             containsKeyIntLike | NuGet50 |            NuGet50 |  100 |    257.99 ns |     3.747 ns |     0.973 ns |  4.32 |    0.03 |    3 | 0.0526 |     - |     - |     331 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|              containsKeyString |   After |              After |  100 |     53.97 ns |     2.617 ns |     0.680 ns |  1.00 |    0.00 |    2 |      - |     - |     - |         - |
|              containsKeyString |  Main50 |             Main50 |  100 |     43.48 ns |     2.182 ns |     0.567 ns |  0.81 |    0.01 |    1 |      - |     - |     - |         - |
|              containsKeyString | NuGet50 |            NuGet50 |  100 |     73.30 ns |     2.800 ns |     0.727 ns |  1.36 |    0.01 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|             containsKeyRefLike |   After |              After |  100 |    146.72 ns |    10.741 ns |     2.789 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|             containsKeyRefLike |  Main50 |             Main50 |  100 |    229.40 ns |     5.506 ns |     0.852 ns |  1.56 |    0.03 |    2 |      - |     - |     - |         - |
|             containsKeyRefLike | NuGet50 |            NuGet50 |  100 |    249.09 ns |    10.063 ns |     1.557 ns |  1.69 |    0.03 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|           containsKeyIntRecord |   After |              After |  100 |    102.18 ns |     2.399 ns |     0.623 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|           containsKeyIntRecord |  Main50 |             Main50 |  100 |    233.32 ns |     8.690 ns |     1.345 ns |  2.28 |    0.03 |    2 | 0.0526 |     - |     - |     331 B |
|           containsKeyIntRecord | NuGet50 |            NuGet50 |  100 |    237.66 ns |     8.665 ns |     2.250 ns |  2.33 |    0.01 |    2 | 0.0519 |     - |     - |     331 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|        containsKeyIntRefRecord |   After |              After |  100 |    153.52 ns |     1.265 ns |     0.329 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|        containsKeyIntRefRecord |  Main50 |             Main50 |  100 |    204.53 ns |     6.879 ns |     1.064 ns |  1.33 |    0.01 |    2 |      - |     - |     - |         - |
|        containsKeyIntRefRecord | NuGet50 |            NuGet50 |  100 |    218.81 ns |     4.612 ns |     1.198 ns |  1.43 |    0.00 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|    containsKeyIntLikeNonGenCmp |   After |              After |  100 |     94.30 ns |     3.296 ns |     0.856 ns |  1.00 |    0.00 |    1 | 0.0262 |     - |     - |     166 B |
|    containsKeyIntLikeNonGenCmp |  Main50 |             Main50 |  100 |    238.60 ns |     3.182 ns |     0.826 ns |  2.53 |    0.02 |    2 | 0.0524 |     - |     - |     331 B |
|    containsKeyIntLikeNonGenCmp | NuGet50 |            NuGet50 |  100 |    240.68 ns |    13.277 ns |     2.055 ns |  2.56 |    0.03 |    2 | 0.0521 |     - |     - |     331 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|    containsKeyRefLikeNonGenCmp |   After |              After |  100 |    182.53 ns |     1.547 ns |     0.239 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|    containsKeyRefLikeNonGenCmp |  Main50 |             Main50 |  100 |    231.61 ns |     5.116 ns |     0.792 ns |  1.27 |    0.00 |    2 |      - |     - |     - |         - |
|    containsKeyRefLikeNonGenCmp | NuGet50 |            NuGet50 |  100 |    259.06 ns |    30.456 ns |     4.713 ns |  1.42 |    0.02 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|                   itemCountInt |   After |              After |  100 |    246.34 ns |     6.113 ns |     1.588 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|                   itemCountInt |  Main50 |             Main50 |  100 |    245.57 ns |     9.633 ns |     1.491 ns |  1.00 |    0.01 |    1 |      - |     - |     - |         - |
|                   itemCountInt | NuGet50 |            NuGet50 |  100 |    504.93 ns |    23.024 ns |     5.979 ns |  2.05 |    0.02 |    2 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|               itemCountIntLike |   After |              After |  100 |    248.94 ns |    11.787 ns |     3.061 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|               itemCountIntLike |  Main50 |             Main50 |  100 |    247.55 ns |     3.088 ns |     0.802 ns |  0.99 |    0.01 |    1 |      - |     - |     - |         - |
|               itemCountIntLike | NuGet50 |            NuGet50 |  100 |    511.33 ns |    24.590 ns |     6.386 ns |  2.05 |    0.03 |    2 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|                itemCountString |   After |              After |  100 |    220.37 ns |     5.541 ns |     1.439 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|                itemCountString |  Main50 |             Main50 |  100 |    275.03 ns |   101.002 ns |    26.230 ns |  1.25 |    0.13 |    2 |      - |     - |     - |         - |
|                itemCountString | NuGet50 |            NuGet50 |  100 |    656.19 ns |    39.166 ns |    10.171 ns |  2.98 |    0.06 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|               itemCountRefLike |   After |              After |  100 |    223.71 ns |    10.711 ns |     1.658 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|               itemCountRefLike |  Main50 |             Main50 |  100 |    232.75 ns |    44.043 ns |     6.816 ns |  1.04 |    0.03 |    2 |      - |     - |     - |         - |
|               itemCountRefLike | NuGet50 |            NuGet50 |  100 |    597.86 ns |     7.059 ns |     1.092 ns |  2.67 |    0.02 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|             itemCountIntRecord |   After |              After |  100 |    246.47 ns |     4.839 ns |     0.749 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|             itemCountIntRecord |  Main50 |             Main50 |  100 |    251.68 ns |     5.412 ns |     1.405 ns |  1.02 |    0.01 |    2 |      - |     - |     - |         - |
|             itemCountIntRecord | NuGet50 |            NuGet50 |  100 |    512.38 ns |    16.553 ns |     4.299 ns |  2.07 |    0.02 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|          itemCountIntRefRecord |   After |              After |  100 |    223.56 ns |     7.727 ns |     2.007 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|          itemCountIntRefRecord |  Main50 |             Main50 |  100 |    234.66 ns |    23.713 ns |     6.158 ns |  1.05 |    0.03 |    2 |      - |     - |     - |         - |
|          itemCountIntRefRecord | NuGet50 |            NuGet50 |  100 |    590.23 ns |    11.160 ns |     2.898 ns |  2.64 |    0.03 |    3 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|      itemCountIntLikeNonGenCmp |   After |              After |  100 |    246.00 ns |     3.057 ns |     0.794 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|      itemCountIntLikeNonGenCmp |  Main50 |             Main50 |  100 |    252.91 ns |    33.318 ns |     5.156 ns |  1.03 |    0.02 |    1 |      - |     - |     - |         - |
|      itemCountIntLikeNonGenCmp | NuGet50 |            NuGet50 |  100 |    508.31 ns |    14.288 ns |     3.710 ns |  2.07 |    0.01 |    2 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|      itemCountRefLikeNonGenCmp |   After |              After |  100 |    227.10 ns |    25.343 ns |     3.922 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|      itemCountRefLikeNonGenCmp |  Main50 |             Main50 |  100 |    229.97 ns |    17.960 ns |     4.664 ns |  1.01 |    0.01 |    1 |      - |     - |     - |         - |
|      itemCountRefLikeNonGenCmp | NuGet50 |            NuGet50 |  100 |    589.75 ns |    20.821 ns |     3.222 ns |  2.60 |    0.03 |    2 |      - |     - |     - |         - |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|              maximumElementInt |   After |              After |  100 |     17.88 ns |     0.358 ns |     0.055 ns |  1.00 |    0.00 |    1 | 0.0042 |     - |     - |      26 B |
|              maximumElementInt |  Main50 |             Main50 |  100 |     18.23 ns |     0.739 ns |     0.114 ns |  1.02 |    0.00 |    2 | 0.0041 |     - |     - |      26 B |
|              maximumElementInt | NuGet50 |            NuGet50 |  100 |     40.90 ns |     1.787 ns |     0.464 ns |  2.28 |    0.02 |    3 | 0.0041 |     - |     - |      26 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|          maximumElementIntLike |   After |              After |  100 |     18.44 ns |     0.296 ns |     0.077 ns |  1.00 |    0.00 |    1 | 0.0042 |     - |     - |      26 B |
|          maximumElementIntLike |  Main50 |             Main50 |  100 |     19.00 ns |     1.713 ns |     0.265 ns |  1.03 |    0.01 |    2 | 0.0041 |     - |     - |      26 B |
|          maximumElementIntLike | NuGet50 |            NuGet50 |  100 |     40.59 ns |     1.399 ns |     0.363 ns |  2.20 |    0.02 |    3 | 0.0041 |     - |     - |      26 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|           maximumElementString |   After |              After |  100 |     27.09 ns |     0.640 ns |     0.166 ns |  1.00 |    0.00 |    1 | 0.0041 |     - |     - |      26 B |
|           maximumElementString |  Main50 |             Main50 |  100 |     28.31 ns |     0.553 ns |     0.086 ns |  1.04 |    0.01 |    2 | 0.0041 |     - |     - |      26 B |
|           maximumElementString | NuGet50 |            NuGet50 |  100 |     61.37 ns |     0.794 ns |     0.123 ns |  2.26 |    0.01 |    3 | 0.0040 |     - |     - |      26 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|          maximumElementRefLike |   After |              After |  100 |     26.51 ns |     1.453 ns |     0.377 ns |  1.00 |    0.00 |    1 | 0.0041 |     - |     - |      26 B |
|          maximumElementRefLike |  Main50 |             Main50 |  100 |     27.01 ns |     0.422 ns |     0.110 ns |  1.02 |    0.02 |    1 | 0.0041 |     - |     - |      26 B |
|          maximumElementRefLike | NuGet50 |            NuGet50 |  100 |     55.84 ns |     3.147 ns |     0.817 ns |  2.11 |    0.05 |    2 | 0.0041 |     - |     - |      26 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|        maximumElementIntRecord |   After |              After |  100 |     19.60 ns |     1.537 ns |     0.238 ns |  1.00 |    0.00 |    1 | 0.0042 |     - |     - |      26 B |
|        maximumElementIntRecord |  Main50 |             Main50 |  100 |     20.40 ns |     0.529 ns |     0.137 ns |  1.04 |    0.01 |    2 | 0.0042 |     - |     - |      26 B |
|        maximumElementIntRecord | NuGet50 |            NuGet50 |  100 |     42.66 ns |     0.841 ns |     0.218 ns |  2.18 |    0.02 |    3 | 0.0040 |     - |     - |      26 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|     maximumElementIntRefRecord |   After |              After |  100 |     25.98 ns |     0.350 ns |     0.091 ns |  1.00 |    0.00 |    1 | 0.0042 |     - |     - |      26 B |
|     maximumElementIntRefRecord |  Main50 |             Main50 |  100 |     27.13 ns |     1.662 ns |     0.432 ns |  1.04 |    0.02 |    2 | 0.0042 |     - |     - |      26 B |
|     maximumElementIntRefRecord | NuGet50 |            NuGet50 |  100 |     55.47 ns |     3.145 ns |     0.817 ns |  2.14 |    0.04 |    3 | 0.0042 |     - |     - |      26 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
| maximumElementIntLikeNonGenCmp |   After |              After |  100 |     18.59 ns |     0.261 ns |     0.068 ns |  1.00 |    0.00 |    1 | 0.0041 |     - |     - |      26 B |
| maximumElementIntLikeNonGenCmp |  Main50 |             Main50 |  100 |     18.78 ns |     1.821 ns |     0.473 ns |  1.01 |    0.02 |    1 | 0.0041 |     - |     - |      26 B |
| maximumElementIntLikeNonGenCmp | NuGet50 |            NuGet50 |  100 |     40.25 ns |     0.616 ns |     0.160 ns |  2.17 |    0.01 |    2 | 0.0040 |     - |     - |      26 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
| maximumElementRefLikeNonGenCmp |   After |              After |  100 |     25.94 ns |     0.787 ns |     0.122 ns |  1.00 |    0.00 |    1 | 0.0042 |     - |     - |      26 B |
| maximumElementRefLikeNonGenCmp |  Main50 |             Main50 |  100 |     26.39 ns |     1.779 ns |     0.275 ns |  1.02 |    0.01 |    2 | 0.0041 |     - |     - |      26 B |
| maximumElementRefLikeNonGenCmp | NuGet50 |            NuGet50 |  100 |     55.08 ns |     1.256 ns |     0.326 ns |  2.12 |    0.02 |    3 | 0.0042 |     - |     - |      26 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|                      subsetInt |   After |              After |  100 |  2,051.29 ns |    35.383 ns |     5.475 ns |  1.00 |    0.00 |    1 |      - |     - |     - |      24 B |
|                      subsetInt |  Main50 |             Main50 |  100 |  3,498.46 ns |   372.447 ns |    96.723 ns |  1.70 |    0.05 |    2 |      - |     - |     - |      32 B |
|                      subsetInt | NuGet50 |            NuGet50 |  100 |  5,202.80 ns |   254.615 ns |    39.402 ns |  2.54 |    0.02 |    3 |      - |     - |     - |      32 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|                  subsetIntLike |   After |              After |  100 |  6,212.42 ns |   118.741 ns |    18.375 ns |  1.00 |    0.00 |    1 |      - |     - |     - |      24 B |
|                  subsetIntLike |  Main50 |             Main50 |  100 | 23,036.75 ns |   762.015 ns |   117.923 ns |  3.71 |    0.02 |    2 | 4.4643 |     - |     - |   28304 B |
|                  subsetIntLike | NuGet50 |            NuGet50 |  100 | 23,672.89 ns | 1,304.249 ns |   201.834 ns |  3.81 |    0.03 |    3 | 4.4643 |     - |     - |   28304 B |
|                                |         |                    |      |              |              |              |       |         |      |        |       |       |           |
|                   subsetString |   After |              After |  100 |  7,616.72 ns |   670.060 ns |   174.012 ns |  1.00 |    0.00 |    2 |      - |     - |     - |      24 B |
|                   subsetString |  Main50 |             Main50 |  100 |  6,297.17 ns |   572.451 ns |    88.587 ns |  0.83 |    0.02 |    1 |      - |     - |     - |      32 B |
|                   subsetString | NuGet50 |            NuGet50 |  100 | 11,107.61 ns |   679.342 ns |   176.423 ns |  1.46 |    0.04 |    3 |      - |     - |     - |      32 B |
…