@buybackoff 2021-01-10
Cannot yet understand why the build fails or reproduce locally with meaningful output.

Locally all but 10 tests from FSharp.Core do pass.

10 tests, e.g. `_.``Uint16s.Collection.ArrayArray |> Array.map Set.ofArray C.I.compare``()` fail with what seems like an expectation that a compare function returns only `[-1,0,1]` instead of `[negative,0,positive]` values. The failure happens only for small types, for which comparer is implemented by cast to int and sub: `int(l:int16) - int(r:int16)`, this never overflows for int types <= 16 bytes => it's safe.

Does F# spec require `[-1,0,1]` or the tests depend on an implementation detail?

![image](https://user-images.githubusercontent.com/489852/104119804-77e68b80-5332-11eb-8aa3-354bbc564721.png)