@cartermp 2021-01-12
Re-triggering CI to see if there's something wrong environmentally, every CI leg shows an NRE when building the boostrap compiler which would be unrelated to your changes...