@cartermp 2021-01-12
@buybackoff yes, there are I think a few tests that _I think_ do depend on some implementation details, in comparersregression.fs. It's a ridiculously large file that's generated by a tool, and consequently, kind of hard to diagnose :(

@dsyme could speak more towards the F# spec in this space.