@buybackoff 2021-01-12
> yes, there are I think a few tests that I think do depend on some implementation details, in comparersregression.fs. It's a ridiculously large file that's generated by a tool, and consequently, kind of hard to diagnose :(

In the last commit I've forced Set.compare to return [-1;0;1] and all tests for FSharp.Core pass locally

> Re-triggering CI to see if there's something wrong environmentally, every CI leg shows an NRE when building the boostrap compiler which would be unrelated to your changes...

It is related. I isolated the cause to these lines:

https://github.com/dotnet/fsharp/blob/0900934739d8194c42b967f5cb46f7a759c89187/src/fsharp/FSharp.Core/mapsetcmp.fs#L26-L30

If I add `&& normalCmp` to the `IComparable<T>` `if` check below, everything works. There are two possibilities:
* There is a type that implements `IComparable<T>` and has structural comparison, but the two work differently for `T`
* Some code depends on `[-1;0;1]` return values (but I tried to build after making `Set.compare` to return `[-1;0;1]` but it still fails. I've also tried to exlude all types of tuples as in #9348)