@cartermp 2021-01-23
Will close this out as per https://github.com/dotnet/fsharp/pull/10855#issuecomment-765709617

(in case folks don't click the link, the rationale is that regressing reference types here could make the tradeoff not worth it, at least not without more information that it's the right tradeoff.)